requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1

# PDF parsing
PyMuPDF==1.24.4

# Embeddings
sentence-transformers==3.0.1
torch==2.3.1          # CPU build; remove +cpu suffix if you have GPU

# Vector store
faiss-cpu==1.8.0      # use faiss-gpu if you have CUDA

# LLM
groq==0.9.0

# Frontend
streamlit==1.35.0
requests==2.32.3

# Utilities
numpy==1.26.4